When find placeholder for element in window prefer one that is to be rendered
For this bug to occur the method ModelServiceImpl.findPlaceholderFor(MWindow, MUIElement)
is called when a dropped view part is (re-)activated via
  {@code org.eclipse.e4.ui.workbench.addons.dndaddon.DropAgent.reactivatePart(MUIElement)}.

Currently when searching for a placeholder with the model service for a given element in a given window
- all elements of type MPlaceholder (in a perspective) are returned
- all placeholders referencing the given element are collected in the elementRefs list
- if there is more than one placeholder and none is contained in the "shared area"
--> the first placeholder in the elementRefs list is returned.

When following the steps to reproduce in any case a placeholder that was _not_ to be rendered was returned,
activated and visible. The view that was dropped into another part stack was also visible.

As written in the code comment:

IF the order of the elements is _not_ predefined (somewhere at another place ??) any placeholder could be returned
  and one that is to be rendered should be preferred.

ELSE (the order is predefined somewhere else) the comment "just return the first one" could be changed
to a meaningful comment explaining why the first element has to be returned (or at least given a hint) and
the index (0) could be a constant with a meaningful name.

Fixes #nnnn